% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_colors.R
\name{select_colors}
\alias{select_colors}
\title{Select Colors from Palette}
\usage{
select_colors(palette, num_series)
}
\arguments{
\item{palette}{A list of color palettes generated by \code{color_palette()}.}

\item{num_series}{An integer specifying the number of colors to select.}
}
\value{
A character vector of selected color hex codes.
}
\description{
This function selects a specified number of colors from the
c4cp color palette.
}
\examples{
palettes <- color_palette()
selected <- select_colors(palettes, 10)
plot(x,y, col = selected[1])
}
